requirements.txt
Django==2.2.6
djangorestframework==3.10.3
PyJWT==1.7.1